requirements.txt
openai
chromadb
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
youtube-transcript-api
gradio
python-dotenv
tiktoken
google-api-python-client 
oauth2client
google-auth-httplib2
google-auth-oauthlib